Worked on a Add Workout button
I couldn't figure out how to create a new page so I just modified app.tsx. I tried creating a workout folder and have a workout.tsx inside but that didn't work. The button isn't working right now, the select is going to crash the website. Still working on it.