Created workout page
Other than the obvious cards, the create workout button is the plus sign and it pops up a dialog. The first plus button on the right of workout 1 also pops up a dialog. The problem with the popover for some reason is that it only workout with keyboard and not with mouse.